Very intense update... Writing recursive function to find all the mesh render components in URDF
Very intense update... Writing recursive function to find all the mesh render components in URDF:

This is to allow the application in the future to load custom robots from ROS, and then adapt the color of them. The method is designed to only run the recursive search once, and then store the names in a list to lighten the load.