ActiveRobot and Positioning Update
This includes a minor update that has a big error... Updating Service Manager to set ActiveRobotName to null in order to check it against the incoming message. The second Major update it includes changing the order of robot positioning on trajectory instantiation in order to make sure it is under the correct parent before changing the local position.