Update robot_base_frame parsing.
This is due to the structure of the in comming message changing. The primary difference is that the message now Inherits from dict, which means that the structure of the message follows the compasObject.__data__ structure, and sends the messages in an extended dictionary format. This is a small change, but caused major problems with the parsing of the robot_base_frame.